bookyourticket/requirements.txt
Django==2.2
django-crispy-forms==1.7.2
djangorestframework==3.9.4
gunicorn
pillow
